ref(trace): refactor
