+++ static/app/views/performance/newTraceDetails/traceDrawer/details/transaction.tsx
@@ -54,7 +54,7 @@ import {transactionSummaryRouteWithQuery} from 'sentry/views/performance/transac
 
 import type {TraceTree, TraceTreeNode} from '../../traceTree';
 
-import IssueList from './issues/issueList';
+import {IssueList} from './issues/issues';
 import {TraceDrawerComponents} from './styles';
 
 function OpsBreakdown({event}: {event: EventTransaction}) {
@@ -309,14 +309,14 @@ export function TransactionNodeDetails({
         </TraceDrawerComponents.Actions>
       </TraceDrawerComponents.HeaderContainer>
 
-      {hasIssues && (
+      {hasIssues ? (
         <IssueList
-          nodeType="transaction"
+          node={node}
           organization={organization}
           issues={relatedIssues}
           event_id={event.id}
         />
-      )}
+      ) : null}
 
       <TraceDrawerComponents.Table className="table key-value">
         <tbody>
